% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{calculateBasis}
\alias{calculateBasis}
\title{Calculate Basis functions for new observations}
\usage{
calculateBasis(z, nIMax, system)
}
\arguments{
\item{z}{elements where basis is going to be calculated}

\item{nIMax}{how many functions should be calculated}

\item{system}{Basis system to be used. Currently, either "cosine" or "Fourier"}
}
\value{
Test
}
\description{
Calculate Basis functions for new observations
}

