% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{cde_loss}
\alias{cde_loss}
\title{Calculate CDE loss}
\usage{
cde_loss(pred, z_grid, z_test)
}
\arguments{
\item{pred}{a matrix of conditional density estimates; rows
correspond to densities on z_grid and columns correspond to
observations z_test.}

\item{z_grid}{a vector of grid points at which pred is evaluated.}

\item{z_test}{a vector of observed z-values.}
}
\value{
The estimated CDE loss for the predict estimates
}
\description{
Calculate CDE loss
}
