% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{chooseDelta}
\alias{chooseDelta}
\title{Title Choose threshold value}
\usage{
chooseDelta(objectCDE, xValidation, zValidation, deltaGrid = seq(0, 0.4,
  0.05))
}
\arguments{
\item{objectCDE}{An object of the class FlexCoDE with a fitted CDE, typically fitted used fitFlexCoDE beforehand}

\item{xValidation}{Covariates x used to validate (tune) the model (one x observation per row).}

\item{zValidation}{Responses z used to validate (tune) the model  (matrix with 1 column). Each row corresponds to a row of the xValidation argument}
}
\value{
Best Delta
}
\description{
Title Choose threshold value
}

