% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctionsKernel.R
\name{chooseDeltaKernel}
\alias{chooseDeltaKernel}
\title{Choose threshold value to remove spurius bumps}
\usage{
chooseDeltaKernel(
  objectCDE,
  kernelValidationTrain,
  zValidation,
  deltaGrid = seq(0, 0.4, 0.05)
)
}
\arguments{
\item{objectCDE}{An object of the class FlexCoDE with a fitted CDE, typically fitted used \code{\link{fitFlexCoDE}} beforehand}

\item{zValidation}{Responses z used to validate (tune) the model  (matrix with 1 column). Each row corresponds to a row of the xValidation argument}

\item{xValidation}{Covariates x used to validate (tune) the model (one x observation per row).}
}
\value{
Best delta
}
\description{
This function is typically not directly used by the user; it is used inside  \code{\link{fitFlexCoDE}}
}
