% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{chooseSharpen}
\alias{chooseSharpen}
\title{Choose sharpen parameter of a conditional density estimator}
\usage{
chooseSharpen(
  objectCDE,
  X,
  Z,
  alpha_grid = seq(0.01, 10, length.out = 20),
  n_grid = n_grid
)
}
\arguments{
\item{objectCDE}{An object of the class FlexCoDE with a fitted CDE,
typically fitted used \code{\link{fitFlexCoDE}} beforehand}

\item{X}{Covariates used to validate (tune) the model (one x
observation per row).}

\item{Z}{Responses used to validate (tune) the model (matrix with 1
column). Each row corresponds to a row of the xValidation
argument}

\item{alpha_grid}{Grid of values of alpha}
}
\value{
Value of alpha which minimizes CDE loss
}
\description{
Chooses optimal alpha for the new conditional density
estimator f(z|x) <- f^alpha(z|x).
}
