% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctionsKernel.R
\name{combineFlexCoDEKernel}
\alias{combineFlexCoDEKernel}
\title{Finds best linear combination of several FlexCoDE estimates}
\usage{
combineFlexCoDEKernel(objectCDE_binded, kernelValidationTrain, zValidation,
  kernelTestTrain = NULL, zTest = NULL)
}
\arguments{
\item{objectCDE_binded}{An object of the class FlexCoDE_binded with a fitted CDE, typically fitted used \code{\link{bindFlexCoDE}} beforehand}

\item{zValidation}{Responses z used to validate (tune) the model  (matrix with 1 column). Each row corresponds to a row of the xValidation argument}

\item{zTest}{Responses z used to estimate risk of final model  (matrix with one column; one observation per row). Default is NULL}

\item{xValidation}{Covariates x used to validate (tune) the model (one x observation per row).}

\item{xTest}{Covariates x used to estimate risk of final model (one observation per row; same number of columns as xTrain). Default is NULL}
}
\value{
Returns an object of the class "combinedFlexCoDE" which contains the weights best linear combination of the input models, together with all fitted models
}
\description{
Finds best linear combination of several FlexCoDE estimates
}
