% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions (Cópia em conflito de rizbicki-Gazelle-Professional 2017-10-04).R, R/auxiliaryFunctions.R
\name{cosine_basis}
\alias{cosine_basis}
\title{Evaluates cosine basis for new observations}
\usage{
cosine_basis(z, n_basis)

cosine_basis(z, n_basis)
}
\arguments{
\item{z}{elements where basis is going to be calculated. Assumes z
has been scaled to [0, 1]}

\item{n_basis}{how many basis functions should be calculated}
}
\value{
A matrix of dimension length(z) by n_basis with entries
  consisting of the first n_basis cosine basis functions evaluated
  at the points z

A matrix of dimension length(z) by n_basis with entries
  consisting of the first n_basis cosine basis functions evaluated
  at the points z
}
\description{
Evaluates cosine basis for new observations

Evaluates cosine basis for new observations
}

