% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions (Cópia em conflito de rizbicki-Gazelle-Professional 2017-10-04).R, R/basicFunctions.R
\name{estimateError}
\alias{estimateError}
\title{Estimate error (risk) of FlexCoDE object via test set}
\usage{
estimateError(obj, x_test, z_test, se = TRUE, n_boot = 500, n_grid = 1000)

estimateError(obj, x_test, z_test, se = TRUE, n_boot = 500, n_grid = 1000)
}
\arguments{
\item{obj}{is an object of the class FlexCoDE typically fitted used
\code{\link{fitFlexCoDE}}}

\item{x_test}{Covariates for the sample used to test the model (one
observation per row)}

\item{z_test}{Response for the sample used to test the model (one
observation per row)}

\item{se}{Boolean flag determining if bootstrap standard error are
computed. Default is TRUE}

\item{n_boot}{Number of bootstrap samples used to calculate
standard errors. Default is 500}

\item{n_grid}{Number of grid points to evaluate conditional
density. Default is 500.}

\item{obj}{is an object of the class FlexCoDE typically fitted used
\code{\link{fitFlexCoDE}}}

\item{x_test}{Covariates for the sample used to test the model (one
observation per row)}

\item{z_test}{Response for the sample used to test the model (one
observation per row)}

\item{se}{Boolean flag determining if bootstrap standard error are
computed. Default is TRUE}

\item{n_boot}{Number of bootstrap samples used to calculate
standard errors. Default is 500}

\item{n_grid}{Number of grid points to evaluate conditional
density. Default is 500.}
}
\value{
Estimated error (with SE if se = TRUE)

Estimated error (with SE if se = TRUE)
}
\description{
Estimate error (risk) of FlexCoDE object via test set

Estimate error (risk) of FlexCoDE object via test set
}

