% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{estimateErrorFlexCoDE}
\alias{estimateErrorFlexCoDE}
\title{Estimate error (risk) of FlexCoDE object via test set}
\usage{
estimateErrorFlexCoDE(objectCDE = objectCDE, xTest, zTest, se = TRUE)
}
\arguments{
\item{objectCDE}{is an object of the class FlexCoDEtypically typically fitted used \code{\link{fitFlexCoDE}} beforehand}

\item{xTest}{Covariates x of the sample used to test the model (one observation per row)}

\item{zTest}{Response z of the sample used to test the model (one observation per row)}

\item{se}{Should standard error be computed? Default is TRUE}
}
\value{
Estimated error (with SE if desired)
}
\description{
Estimate error (risk) of FlexCoDE object via test set
}

