% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctionsKernel.R
\name{estimateErrorFlexCoDEKernel}
\alias{estimateErrorFlexCoDEKernel}
\title{Estimate error (risk) of FlexCoDE object via test set}
\usage{
estimateErrorFlexCoDEKernel(objectCDE, kernelTestTrain, zTest, se = TRUE)
}
\arguments{
\item{objectCDE}{is an object of the class FlexCoDEtypically typically fitted used \code{\link{fitFlexCoDE}} beforehand}

\item{zTest}{Response z of the sample used to test the model (one observation per row)}

\item{se}{Should standard error be computed? Default is TRUE}

\item{xTest}{Covariates x of the sample used to test the model (one observation per row)}
}
\value{
Estimated error (with SE if desired)
}
\description{
Estimate error (risk) of FlexCoDE object via test set
}
