% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{fitFlexCoDE}
\alias{fitFlexCoDE}
\title{FlexCoDE Fit Conditional Density Estimation via Regression}
\usage{
fitFlexCoDE(xTrain, zTrain, xValidation, zValidation, xTest = NULL,
  zTest = NULL, nIMax = length(zTrain), regressionFunction,
  regressionFunction.extra = NULL, system = "Fourier", deltaGrid = seq(0,
  0.4, 0.05), chooseDelta = TRUE, verbose = TRUE)
}
\arguments{
\item{xTrain}{Covariates x used to train the model (one observation per row)}

\item{zTrain}{Responses z used to train the model  (matrix with one column; one observation per row)}

\item{xValidation}{Covariates x used to tune the model (one observation per row; same number of columns as xTrain)}

\item{zValidation}{Responses z used to tune the model  (matrix with one column; one observation per row)}

\item{xTest}{Covariates x used to estimate risk of final model (one observation per row; same number of columns as xTrain). Default is NULL}

\item{zTest}{Responses z used to estimate risk of final model  (matrix with one column; one observation per row). Default is NULL}

\item{nIMax}{Maximum possible number of components of the series expansion (that is, the function will find the best I<nIMax). Default is 100}

\item{regressionFunction}{a function indicating which regression method will be used to estimate the expansion coefficients. Currently can be one of}

\item{regressionFunction.extra}{extra parameters to be sent to regression function; see the regression you want to use to check what are the available parameters}

\item{system}{Basis for z. Current options are "Fourier", "Cosine" and "discrete". Default is "Fourier"}

\item{deltaGrid}{Grid of threshold deltas (betwen 0 and 0.5). Default value is seq(0,0.4,0.05).}

\item{chooseDelta}{Should delta, the cutoff to remove spurious bumps, be chosen?}

\item{verbose}{Should we print what we are doing? Default is TRUE.}
}
\value{
Returns the fitted estimated conditional density, and object of the class FlexCoDE
}
\description{
FlexCoDE Fit Conditional Density Estimation via Regression
}

