% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{FlexZBoost}
\alias{FlexZBoost}
\title{Fits FlexZBoost}
\usage{
FlexZBoost(xTrain, zTrain, xValidation, zValidation, xTest = NULL,
  zTest = NULL, chooseSharpen = TRUE, ...)
}
\arguments{
\item{xTrain}{Covariates x used to train the model (one observation per row)}

\item{zTrain}{Responses z used to train the model  (matrix with one column; one observation per row)}

\item{xValidation}{Covariates x used to tune the model (one observation per row; same number of columns as xTrain)}

\item{zValidation}{Responses z used to tune the model  (matrix with one column; one observation per row)}

\item{xTest}{Covariates x used to estimate risk of final model (one observation per row; same number of columns as xTrain). Default is NULL}

\item{zTest}{Responses z used to estimate risk of final model  (matrix with one column; one observation per row). Default is NULL}

\item{chooseSharpen}{Should alpha, the parameter to sharpen the final estimate, be chosen? Default is TRUE}

\item{...}{additional arguments to fitFlexCoDE}
}
\value{
Returns the fitted estimated conditional density, and object of the class FlexCoDE. The return value is an object with the following components:
\item{zMin, zMax}{Minimum and maximum value of z}
\item{nIMax}{Maximum number of expansion coefficients (user input). Default is minimum between 25 and number of training samples.}
\item{system}{Basis used for expanding the response}
\item{zTrain}{zTrain (user input)}
\item{xTrain}{xTrain (user input)}
\item{regressionObject}{Object with fitted regressions. Class and content depend on which regression method was chosen by user}
\item{errors}{Estimated errors for each value of I (number of expansion coefficients) using validation set}
\item{bestI}{Optimal number of I according to validation set}
\item{bestError}{Estimated error of model with bestI expansion terms according to validation set}
\item{bestDelta}{Optimal value of threshold delta according to validation set}
\item{bestAlpha}{Optimal value of alpha according to validation set}
\item{estimatedRisk}{(If user provides xTest and zTest) Estimated risk (error) according to test set)}
}
\description{
Wrapper for fitFlexCoDE for the special case of FlexZBoost
}
