% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{haar_basis}
\alias{haar_basis}
\title{Evaluates Haar basis for new observations}
\usage{
haar_basis(z, n_basis)
}
\arguments{
\item{z}{elements where basis is going to be calculated. Assumes z
has been scaled to [0, 1]}

\item{n_basis}{how many basis functions should be calculated}
}
\value{
A matrix of dimension length(z) by n_basis with entries
  consisting of the first n_basis Haar basis functions evaluated
  at the points z
}
\description{
Evaluates Haar basis for new observations
}

