% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{normalize_density}
\alias{normalize_density}
\title{Project onto density estimates
Uses binary search to determine the correct cutoff}
\usage{
normalize_density(bin_size, estimates, tolerance = 0.001, max_iters = 500)
}
\description{
Project onto density estimates
Uses binary search to determine the correct cutoff
}
