% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions (Cópia em conflito de rizbicki-Gazelle-Professional 2017-10-04).R, R/basicFunctions.R
\name{plot.FlexCoDE_binded}
\alias{plot.FlexCoDE_binded}
\title{Plots examples of estimated densities together with real response}
\usage{
plot.FlexCoDE_binded(objectCDE_binded, xTest, zTest, nPlots = min(nrow(xTest),
  9), fontSize = 12, lineWidth = 1)

plot.FlexCoDE_binded(objectCDE_binded, xTest, zTest, nPlots = min(nrow(xTest),
  9), fontSize = 12, lineWidth = 1)
}
\arguments{
\item{objectCDE_binded}{Object of the class "FlexCoDE_binded", typically obtained using \code{\link{bindFlexCoDE}} beforehand}

\item{xTest}{Covariates x of the sample used to test the model (one observation per row)}

\item{zTest}{Response z of the sample used to test the model (one observation per row)}

\item{nPlots}{Number of desired densities to be ploted (which will be picked at random). Default is minimum between 8 and number of testing points}

\item{fontSize}{Font size of axis labels and legend}

\item{lineWidth}{Line width of the curves to be ploted}

\item{objectCDE_binded}{Object of the class "FlexCoDE_binded", typically obtained using \code{\link{bindFlexCoDE}} beforehand}

\item{xTest}{Covariates x of the sample used to test the model (one observation per row)}

\item{zTest}{Response z of the sample used to test the model (one observation per row)}

\item{nPlots}{Number of desired densities to be ploted (which will be picked at random). Default is minimum between 8 and number of testing points}

\item{fontSize}{Font size of axis labels and legend}

\item{lineWidth}{Line width of the curves to be ploted}
}
\value{
Plot with estimated densities

Plot with estimated densities
}
\description{
Plots examples of estimated densities together with real response

Plots examples of estimated densities together with real response
}
\examples{
# See \\code{\\link{bindFlexCoDE}}

# See \\code{\\link{bindFlexCoDE}}

}

