% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctionsKernel.R
\name{plot.FlexCoDE_bindedKernel}
\alias{plot.FlexCoDE_bindedKernel}
\title{Plots examples of estimated densities together with real response}
\usage{
\method{plot}{FlexCoDE_bindedKernel}(
  objectCDE_binded,
  kernelTestTrain,
  zTest,
  nPlots = min(nrow(kernelTestTrain), 8),
  fontSize = 12,
  lineWidth = 1
)
}
\arguments{
\item{objectCDE_binded}{Object of the class "FlexCoDE_binded", typically obtained using \code{\link{bindFlexCoDE}} beforehand}

\item{zTest}{Response z of the sample used to test the model (one observation per row)}

\item{nPlots}{Number of desired densities to be ploted (which will be picked at random). Default is minimum between 8 and number of testing points}

\item{fontSize}{Font size of axis labels and legend}

\item{lineWidth}{Line width of the curves to be ploted}

\item{xTest}{Covariates x of the sample used to test the model (one observation per row)}
}
\value{
Plot with estimated densities
}
\description{
Plots examples of estimated densities together with real response
}
\examples{
# See \code{\link{bindFlexCoDEKernel}}

}
