% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{post_process}
\alias{post_process}
\title{Post process density estimate to normalize and remove bumps}
\usage{
post_process(binSize, estimates, delta = 0, threshold = 1e-06,
  alpha = 1)
}
\arguments{
\item{binSize}{size of grid bins}

\item{estimates}{vector of density estimate on a grid}

\item{delta}{parameter for removing bumps; default is 0.0 implying
no bump removal}

\item{threshold}{parameter for discarding small densities; default
is 1e-6}
}
\value{
A vector of the density estimate
}
\description{
Post process density estimate to normalize and remove bumps
}
