% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctionsKernel.R
\name{predict.combinedFlexCoDEKernel}
\alias{predict.combinedFlexCoDEKernel}
\title{Evaluates the estimated  density of new observations (testing points) of a "combinedFlexCoDE" object}
\usage{
predict.combinedFlexCoDEKernel(objectCombined, kernelNewTrain, B = 1000)
}
\arguments{
\item{B}{Number of point where f(z|x) will be evaluated (on the z scale). This will be equally spaced between zMin and zMax}

\item{objectCDE}{Object of the class "combinedFlexCoDE", typically fitted used \code{\link{combineFlexCoDE}} beforehand}

\item{xNew}{Matrix with nTest rows and same number of columns as xTrain, containing x's for which the estimates are desired.}
}
\value{
The return value is an object with the following components
\item{z}{Points where the density was evaluate}
\item{CDE }{Matrix with value of the density at points z. Each row corresponds to a different observation x (i-th row of CDE corresponds to i-th row of xTest).}
}
\description{
Evaluates the estimated  density of new observations (testing points) of a "combinedFlexCoDE" object
}
\examples{
# See \\code{\\link{combineFlexCoDE}}
}
