% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regressionPredictionFunctions.R
\name{predict.NN}
\alias{predict.NN}
\title{Title}
\usage{
\method{predict}{NN}(object, xNew, maxTerms = NULL)
}
\arguments{
\item{object}{}

\item{xNew}{}

\item{maxTerms}{}
}
\value{
x
}
\description{
Title
}

