% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionPredictionFunctionsKernel.R
\name{predict.SeriesKernel}
\alias{predict.SeriesKernel}
\title{Predict Series regression}
\usage{
predict.SeriesKernel(object, kernelNewTrain, maxTerms = NULL)
}
\arguments{
\item{object}{object of the class Series}

\item{maxTerms}{maximum number of expansion coefficients}

\item{xNew}{matrix with covariates where prediction will be calculated}
}
\value{
returns matrix where element (i,j) contains the estimate of the j-th expansion coefficient for the j-th sample
}
\description{
This function is typically not directly used by the user; it is used inside  \code{\link{fitFlexCoDE}}
}

