% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions (Cópia em conflito de rizbicki-Gazelle-Professional 2017-10-04).R, R/basicFunctions.R
\name{print.combinedFlexCoDE}
\alias{print.combinedFlexCoDE}
\title{Print object of classe combinedFlexCoDE}
\usage{
\method{print}{combinedFlexCoDE}(objectCombined)

\method{print}{combinedFlexCoDE}(objectCombined)
}
\arguments{
\item{objectCDE}{Object of the class "combinedFlexCoDE", typically fitted used \code{\link{combineFlexCoDE}} beforehand}

\item{objectCDE}{Object of the class "combinedFlexCoDE", typically fitted used \code{\link{combineFlexCoDE}} beforehand}
}
\value{
returns information regarding the fitted model

returns information regarding the fitted model
}
\description{
Print object of classe combinedFlexCoDE

Print object of classe combinedFlexCoDE
}

