% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctionsKernel.R
\name{print.combinedFlexCoDEKernel}
\alias{print.combinedFlexCoDEKernel}
\title{Print object of classe combinedFlexCoDE}
\usage{
\method{print}{combinedFlexCoDEKernel}(objectCombined)
}
\arguments{
\item{objectCDE}{Object of the class "combinedFlexCoDE", typically fitted used \code{\link{combineFlexCoDE}} beforehand}
}
\value{
returns information regarding the fitted model
}
\description{
Print object of classe combinedFlexCoDE
}

