% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions (Cópia em conflito de rizbicki-Gazelle-Professional 2017-10-04).R, R/basicFunctions.R
\name{print.FlexCoDE}
\alias{print.FlexCoDE}
\title{Print object of classe FlexCoDE}
\usage{
\method{print}{FlexCoDE}(objectCDE)

\method{print}{FlexCoDE}(objectCDE)
}
\arguments{
\item{objectCDE}{Object of the class "FlexCoDE", typically fitted used \code{\link{fitFlexCoDE}} beforehand}

\item{objectCDE}{Object of the class "FlexCoDE", typically fitted used \code{\link{fitFlexCoDE}} beforehand}
}
\value{
returns information regarding the fitted model

returns information regarding the fitted model
}
\description{
Print object of classe FlexCoDE

Print object of classe FlexCoDE
}

