% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/basicFunctionsKernel.R
\name{print.FlexCoDEKernel}
\alias{print.FlexCoDEKernel}
\title{Print object of classe FlexCoDE}
\usage{
\method{print}{FlexCoDEKernel}(objectCDE)
}
\arguments{
\item{objectCDE}{Object of the class "FlexCoDE", typically fitted used \code{\link{fitFlexCoDE}} beforehand}
}
\value{
returns information regarding the fitted model
}
\description{
Print object of classe FlexCoDE
}

