% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regressionFittingFunctions.R
\name{print.Lasso}
\alias{print.Lasso}
\title{Print function for object of the class Lasso}
\usage{
\method{print}{Lasso}(regressionObject, bestI)
}
\arguments{
\item{regressionObject}{of the class Lasso}

\item{bestI}{optimal number of expansion coefficients}
}
\value{
prints characteristics of the regressions that were fitted
}
\description{
This function is typically not directly used by the user; it is used inside  \code{\link{print.FlexCoDE}}, the print
method for the class FlexCoDE
}

