% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionFittingFunctionsKernel.R
\name{print.NNKernel}
\alias{print.NNKernel}
\title{Print function for object of the class NN}
\usage{
\method{print}{NNKernel}(regressionObject, bestI, nameCovariates)

\method{print}{NNKernel}(regressionObject, bestI, nameCovariates)
}
\arguments{
\item{regressionObject}{of the class NN}

\item{bestI}{optimal number of expansion coefficients}

\item{nameCovariates}{name of the covariates}
}
\value{
prints characteristics of the regressions that were fitted

prints characteristics of the regressions that were fitted
}
\description{
This function is typically not directly used by the user; it is used inside  \code{\link{print.FlexCoDE}}, the print
method for the class FlexCoDE

This function is typically not directly used by the user; it is used inside  \code{\link{print.FlexCoDE}}, the print
method for the class FlexCoDE
}
