% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regressionFittingFunctionsKernel.R
\name{regressionFunction.SDMKernel}
\alias{regressionFunction.SDMKernel}
\title{Title test}
\usage{
regressionFunction.SDMKernel(kernelX, responses, extra = NULL)
}
\arguments{
\item{kernelX}{}

\item{responses}{}

\item{extra}{}
}
\value{
r
}
\description{
Title test
}

