% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionFittingFunctionsKernel.R
\name{regressionFunction.SeriesKernel}
\alias{regressionFunction.SeriesKernel}
\title{Spectral Series Regression}
\usage{
regressionFunction.SeriesKernel(kernelX, responses, extra = NULL)
}
\arguments{
\item{kernelX}{gram matrix that will be used for training}

\item{responses}{matrix where each column is a response for the training data}

\item{extra}{list with one component called nXMax and contains a single integer number that describes what is the maximum number of spectral basis functions with be used}
}
\value{
object of the class Series containing information needed to perform prediction on new points
}
\description{
This function is typically not directly used by the user; it is used inside  \code{\link{fitFlexCoDE}}
}

