% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{remove_bumps}
\alias{remove_bumps}
\title{Removing bumps from density estimates}
\usage{
remove_bumps(binSize, estimates, delta)
}
\arguments{
\item{binSize}{size of grid bins}

\item{delta}{area threshold for removing a bump}

\item{estiamtes}{vector of density estimates on a grid}
}
\value{
A vector of the density with bumps removed
}
\description{
Removing bumps from density estimates
}
