% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{sharpen}
\alias{sharpen}
\title{Sharpen estimated of a conditional density estimator}
\usage{
sharpen(binSize, estimates, alpha)
}
\arguments{
\item{binSize}{size of grid bins}

\item{estimates}{vector of density estimates on a grid}

\item{alpha}{sharpen coefficient}
}
\value{
A vector of the sharp density
}
\description{
Sharpen estimated of a conditional density estimator
}

