% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{bindFlexCoDE}
\alias{bindFlexCoDE}
\title{Binds together objects of the class "FlexCoDE"}
\usage{
bindFlexCoDE(objectCDE1, objectCDE2, ...)
}
\arguments{
\item{objectCDE1}{An object of the class FlexCoDE with a fitted CDE, typically fitted used \code{\link{fitFlexCoDE}} beforehand}

\item{objectCDE2}{An object of the class FlexCoDE with a fitted CDE, typically fitted used \code{\link{fitFlexCoDE}} beforehand}

\item{...}{other objects of the class FlexCoDE with a fitted CDE, typically fitted used \code{\link{fitFlexCoDE}} beforehand}
}
\value{
list with all objects combined. Result is of the class "FlexCoDE_binded"
}
\description{
Binds together objects of the class "FlexCoDE"
}
