% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{calculateBasis}
\alias{calculateBasis}
\title{Calculate Basis functions for new observations}
\usage{
calculateBasis(z, nIMax, system)
}
\arguments{
\item{z}{}

\item{nIMax}{}

\item{system}{}
}
\value{
Test
}
\description{
Calculate Basis functions for new observations
}

