% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{chooseDelta}
\alias{chooseDelta}
\title{Choose threshold value to remove spurius bumps}
\usage{
chooseDelta(objectCDE, xValidation, zValidation, deltaGrid = seq(0, 0.4,
  0.05), nCores = 1)
}
\arguments{
\item{objectCDE}{An object of the class FlexCoDE with a fitted CDE, typically fitted used \code{\link{fitFlexCoDE}} beforehand}

\item{xValidation}{Covariates x used to validate (tune) the model (one x observation per row).}

\item{zValidation}{Responses z used to validate (tune) the model  (matrix with 1 column). Each row corresponds to a row of the xValidation argument}

\item{nCores}{Number of cores to be used for parallel computing. Default is one.}
}
\value{
Best delta
}
\description{
This function is typically not directly used by the user; it is used inside  \code{\link{fitFlexCoDE}}
}

