% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{fitFlexCoDE}
\alias{fitFlexCoDE}
\title{FlexCoDE Fit Conditional Density Estimation via Regression}
\usage{
fitFlexCoDE(xTrain, zTrain, xValidation, zValidation, xTest = NULL,
  zTest = NULL, nIMax = min(25, length(zTrain)), regressionFunction,
  regressionFunction.extra = NULL, system = "Fourier", deltaGrid = seq(0,
  0.45, length.out = 15), chooseDelta = TRUE, verbose = TRUE, zMin = NULL,
  zMax = NULL)
}
\arguments{
\item{xTrain}{Covariates x used to train the model (one observation per row)}

\item{zTrain}{Responses z used to train the model  (matrix with one column; one observation per row)}

\item{xValidation}{Covariates x used to tune the model (one observation per row; same number of columns as xTrain)}

\item{zValidation}{Responses z used to tune the model  (matrix with one column; one observation per row)}

\item{xTest}{Covariates x used to estimate risk of final model (one observation per row; same number of columns as xTrain). Default is NULL}

\item{zTest}{Responses z used to estimate risk of final model  (matrix with one column; one observation per row). Default is NULL}

\item{nIMax}{Maximum possible number of components of the series expansion (that is, the function will find the best I<nIMax). Default is 100}

\item{regressionFunction}{a function indicating which regression method will be used to estimate the expansion coefficients. Currently can be one of}

\item{regressionFunction.extra}{extra parameters to be sent to regression function; see the regression you want to use to check what are the available parameters. The argument nCores which contains the number of cores to be used for parallel computing. Default is one.}

\item{system}{Basis for z. Current options are "Fourier", "Cosine" and "discrete". Default is "Fourier"}

\item{deltaGrid}{Grid of threshold deltas (betwen 0 and 0.5). Default value is seq(0,0.4,0.05).}

\item{chooseDelta}{Should delta, the cutoff to remove spurious bumps, be chosen?}

\item{verbose}{Should we print what we are doing? Default is TRUE.}

\item{zMin}{Minimum value z assumes. Default is min(zTrain).}

\item{zMax}{Maximum value z assumes. Default is max(zTrain).}
}
\value{
Returns the fitted estimated conditional density, and object of the class FlexCoDE. The return value is an object with the following components:
\item{zMin, zMax}{Minimum and maximum value of z}
\item{nIMax}{Maximum number of expansion coefficients (user input). Default is minimum between 25 and number of training samples.}
\item{system}{Basis used for expanding the response}
\item{zTrain}{zTrain (user input)}
\item{xTrain}{xTrain (user input)}
\item{regressionObject}{Object with fitted regressions. Class and content depend on which regression method was chosen by user}
\item{errors}{Estimated errors for each value of I (number of expansion coefficients) using validation set}
\item{bestI}{Optimal number of I according to validation set}
\item{bestError}{Estimated error of model with bestI expansion terms according to validation set}
\item{bestDelta}{Optimal value of threshold delta according to validation set}
\item{estimatedRisk}{(If user provides xTest and zTest) Estimated risk (error) according to test set)}
}
\description{
FlexCoDE Fit Conditional Density Estimation via Regression
}
\examples{
set.seed(400)

# generate data
n=1000
d=10
data=matrix(NA,n,d+1)
data[,1:d]=matrix(rnorm(n*d),n,d)
# generate response from a mixture
data[,d+1]=data[,1]+rnorm(n,0,0.1)


# determine sample sizes
nTrain=round(0.7*n)
nValidation=round(0.25*n)
nTest=n-nTrain-nValidation

# split data
randomIndex=sample(1:n)
xTrain=data[randomIndex[1:nTrain],1:d]
xValidation=data[randomIndex[(nTrain+1):(nTrain+nValidation)],1:d]
xTest=data[randomIndex[(nTrain+nValidation+1):n],1:d]
zTrain=data[randomIndex[1:nTrain],d+1]
zValidation=data[randomIndex[(nTrain+1):(nTrain+nValidation)],d+1]
zTest=data[randomIndex[(nTrain+nValidation+1):n],d+1]

############################################################
###### fit NN estimator (not appropriate for sparse structure)
############################################################
fit=fitFlexCoDE(xTrain,zTrain,xValidation,zValidation,xTest,zTest,
                nIMax = 20,regressionFunction = regressionFunction.NN)
fit$estimatedRisk
print(fit)

# Plot estimated curves evaluates on new test points, and
# compare with true conditional density
predictedValues=predict(fit,xTest,B=500)
plot(predictedValues$z,predictedValues$CDE[1,])
lines(predictedValues$z,dnorm(predictedValues$z,xTest[1,1],0.1),col=2)

# Plot some estimates
plot(fit,xTest,zTest)

# Plot some estimates and prediction regions
plot(fit,xTest,zTest,predictionBandProb = 0.95)


############################################################
# fit SpAM estimator (appropriate for sparse structure); we change
######## default value of tuning parameters
############################################################
fit=fitFlexCoDE(xTrain,zTrain,xValidation,zValidation,xTest,zTest,
                nIMax = 50,regressionFunction = regressionFunction.SpAM,
                regressionFunction.extra=list(sVec=round(seq(1,20,length.out = 5))))
fit$estimatedRisk # much better than NN!
print(fit)


# Plot estimated curves evaluates on new test points, and
# compare with true conditional density
predictedValues=predict(fit,xTest,B=500)
plot(predictedValues$z,predictedValues$CDE[1,])
lines(predictedValues$z,dnorm(predictedValues$z,xTest[1,1],0.1),col=2)


# Same as before, but now using 4 cores for parallel computing
fit=fitFlexCoDE(xTrain,zTrain,xValidation,zValidation,xTest,zTest,
                nIMax = 50,regressionFunction = regressionFunction.SpAM,
                regressionFunction.extra=list(nCores=4,
                                              sVec=round(seq(1,20,length.out = 5))))


############################################################
######## Forest: also appropriate
############################################################
fit=fitFlexCoDE(xTrain,zTrain,xValidation,zValidation,xTest,zTest,
                nIMax = 30,regressionFunction = regressionFunction.Forest)
fit$estimatedRisk # even better than SpAM!
print(fit)

# Plot estimated curves evaluates on new test points, and
# compare with true conditional density
predictedValues=predict(fit,xTest,B=500)
plot(predictedValues$z,predictedValues$CDE[1,])
lines(predictedValues$z,dnorm(predictedValues$z,xTest[1,1],0.1),col=2)

# Plot some estimates and prediction regions
plot(fit,xTest,zTest,predictionBandProb = 0.95)


############################################################
######## Series
############################################################
fit=fitFlexCoDE(xTrain,zTrain,xValidation,zValidation,xTest,zTest,
                nIMax = 20,regressionFunction = regressionFunction.Series)
fit$estimatedRisk
print(fit)


############################################################
######## Lasso
############################################################
fit=fitFlexCoDE(xTrain,zTrain,xValidation,zValidation,xTest,zTest,
                nIMax = 20,regressionFunction = regressionFunction.Lasso)
fit$estimatedRisk
print(fit)

}

