% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctionsKernel.R
\name{fitFlexCoDEKernel}
\alias{fitFlexCoDEKernel}
\title{FlexCoDE Fit Conditional Density Estimation via Regression}
\usage{
fitFlexCoDEKernel(kernelTrainTrain, zTrain, kernelValidationTrain,
  zValidation, kernelTestTrain = NULL, zTest = NULL, nIMax = min(25,
  length(zTrain)), regressionFunction, regressionFunction.extra = NULL,
  system = "Fourier", deltaGrid = seq(0, 0.4, 0.05),
  chooseDelta = TRUE, verbose = TRUE)
}
\arguments{
\item{zTrain}{Responses z used to train the model  (matrix with one column; one observation per row)}

\item{zValidation}{Responses z used to tune the model  (matrix with one column; one observation per row)}

\item{zTest}{Responses z used to estimate risk of final model  (matrix with one column; one observation per row). Default is NULL}

\item{nIMax}{Maximum possible number of components of the series expansion (that is, the function will find the best I<nIMax). Default is 100}

\item{regressionFunction}{a function indicating which regression method will be used to estimate the expansion coefficients. Currently can be one of}

\item{regressionFunction.extra}{extra parameters to be sent to regression function; see the regression you want to use to check what are the available parameters}

\item{system}{Basis for z. Current options are "Fourier", "Cosine" and "discrete". Default is "Fourier"}

\item{deltaGrid}{Grid of threshold deltas (betwen 0 and 0.5). Default value is seq(0,0.4,0.05).}

\item{chooseDelta}{Should delta, the cutoff to remove spurious bumps, be chosen?}

\item{verbose}{Should we print what we are doing? Default is TRUE.}

\item{xTrain}{Covariates x used to train the model (one observation per row)}

\item{xValidation}{Covariates x used to tune the model (one observation per row; same number of columns as xTrain)}

\item{xTest}{Covariates x used to estimate risk of final model (one observation per row; same number of columns as xTrain). Default is NULL}
}
\value{
Returns the fitted estimated conditional density, and object of the class FlexCoDE. The return value is an object with the following components:
\item{zMin, zMax}{Minimum and maximum value of z}
\item{nIMax}{Maximum number of expansion coefficients (user input). Default is minimum between 25 and number of training samples.}
\item{system}{Basis used for expanding the response}
\item{zTrain}{zTrain (user input)}
\item{xTrain}{xTrain (user input)}
\item{regressionObject}{Object with fitted regressions. Class and content depend on which regression method was chosen by user}
\item{errors}{Estimated errors for each value of I (number of expansion coefficients) using validation set}
\item{bestI}{Optimal number of I according to validation set}
\item{bestError}{Estimated error of model with bestI expansion terms according to validation set}
\item{bestDelta}{Optimal value of threshold delta according to validation set}
\item{estimatedRisk}{(If user provides xTest and zTest) Estimated risk (error) according to test set)}
}
\description{
FlexCoDE Fit Conditional Density Estimation via Regression
}
