% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{predict.FlexCoDE}
\alias{predict.FlexCoDE}
\title{Evaluates the estimated  density of new observations (testing points)}
\usage{
\method{predict}{FlexCoDE}(objectCDE, xNew, B = 1000)
}
\arguments{
\item{objectCDE}{Object of the class "FlexCoDE", typically fitted used \code{\link{fitFlexCoDE}} beforehand}

\item{xNew}{Matrix with nTest rows and same number of columns as xTrain, containing x's for which the estimates are desired.}

\item{B}{Number of point where f(z|x) will be evaluated (on the z scale). This will be equally spaced between zMin and zMax}
}
\value{
The return value is an object with the following components
\item{z}{Points where the density was evaluate}
\item{CDE }{Matrix with value of the density at points z. Each row corresponds to a different observation x (i-th row of CDE corresponds to i-th row of xTest).}
}
\description{
Evaluates the estimated  density of new observations (testing points)
}

