% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{predict.FlexCoDE}
\alias{predict.FlexCoDE}
\title{Evaluates the estimated  density of new observations (testing points) of a "FlexCoDE" object}
\usage{
\method{predict}{FlexCoDE}(obj, xNew, B = NULL, predictionBandProb = FALSE, process = TRUE)
}
\arguments{
\item{xNew}{Matrix with nTest rows and same number of columns as xTrain, containing x's for which the estimates are desired.}

\item{B}{Number of point where f(z|x) will be evaluated (on the z scale). This will be equally spaced between zMin and zMax}

\item{predictionBandProb}{Either a number indicating the probability for the highest predictive density region desired  or FALSE if bands are not desired. Default is FALSE}

\item{objectCDE}{Object of the class "FlexCoDE", typically fitted used \code{\link{fitFlexCoDE}} beforehand}
}
\value{
The return value is an object with the following components
\item{z}{Points where the density was evaluate}
\item{CDE}{Matrix with value of the density at points z. Each row corresponds to a different observation x (i-th row of CDE corresponds to i-th row of xTest).}
\item{th}{(If predictionBandProb is not FALSE) Threshold values for each estimated density. The region where estimated densities are above these values have the approximate coverage probability desired. See  \code{\link{plot.FlexCoDE}} for ploting these regions.}
}
\description{
Evaluates the estimated  density of new observations (testing points) of a "FlexCoDE" object
}
