% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionFittingFunctions.R
\name{print.SpAM}
\alias{print.SpAM}
\title{Print function for object of the class SpAM}
\usage{
\method{print}{SpAM}(regressionObject, bestI, nameCovariates)
}
\arguments{
\item{regressionObject}{of the class SpAM}

\item{bestI}{optimal number of expansion coefficients}

\item{nameCovariates}{name of the covariates}
}
\value{
prints characteristics of the regressions that were fitted
}
\description{
This function is typically not directly used by the user; it is used inside  \code{\link{print.FlexCoDE}}, the print
method for the class FlexCoDE
}
