% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionFittingFunctions.R
\name{regressionFunction.Forest}
\alias{regressionFunction.Forest}
\title{Forest Regression}
\usage{
regressionFunction.Forest(x, responses, extra = NULL)
}
\arguments{
\item{x}{matrix with covariates that will be used for training}

\item{responses}{matrix where each column is a response for the training data}

\item{extra}{list with one components named p0Vec, which contains a vetor with different number of variables randomly sampled as candidates at each split of the forest regression (aka mtry in randomForest package); the function will choose the best value among them;  one component named ntree which contains the number of tree to be used by the forest (default is 500), and  one component named maxnodes which contains the number of f terminal nodes trees in the forest can have (if not given, trees are grown to the maximum possible). The list can also contain a component named
nCores which contains the number of cores to be used for parallel computing. Default is one.}
}
\value{
object of the class Forest containing information needed to perform prediction on new points
}
\description{
This function is typically not directly used by the user; it is used inside  \code{\link{fitFlexCoDE}}
}

