% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regressionFittingFunctions.R
\name{regressionFunction.SpAM}
\alias{regressionFunction.SpAM}
\title{Title}
\usage{
regressionFunction.SpAM(x, responses, extra = NULL)
}
\arguments{
\item{x}{}

\item{responses}{}

\item{extra}{}
}
\value{
x
}
\description{
Title
}

