% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionFittingFunctions.R
\name{regressionFunction.SpAM}
\alias{regressionFunction.SpAM}
\title{SpAM Regression (Sparse Additive Model)}
\usage{
regressionFunction.SpAM(x, responses, extra = NULL)
}
\arguments{
\item{x}{matrix with covariates that will be used for training}

\item{responses}{matrix where each column is a response for the training data}

\item{extra}{list with one component named sVec, which contains a vetor with different number of splins; the function will choose the best value among them. The list can also contain a component named
nCores which contains the number of cores to be used for parallel computing. Default is one.}
}
\value{
object of the class SpAM containing information needed to perform prediction on new points
}
\description{
This function is typically not directly used by the user; it is used inside  \code{\link{fitFlexCoDE}}
}

